- change from int to long algorithm DataSet content
git-svn-id: https://svn.code.sf.net/p/auber/code/tulip@1429 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ plugins/colors/EnumColorValuesMapping.cpp
@@ -198,7 +198,7 @@ public:
       composite->addGlEntity(labelVector[i],"label"+sstr.str());
       i++;
     }
-    mainDataSet.set<int>("caption",(int)composite);
+    mainDataSet.set<int>("caption",(long)composite);
     entityDataSet.set<DataSet>("Foreground",mainDataSet);
     dataSet->set<DataSet>("entities",entityDataSet);
